requirements.txt
flask>=3.0.0
selenium>=4.15.0
beautifulsoup4>=4.12.0

lxml>=5.0.0
openpyxl>=3.1.0
requests>=2.31.0
google-genai>=0.1.0
python-dotenv>=1.0.0
